[kmalloc] Make `used_heaps` to be `int`

+++ kernel/kmalloc/general_kmalloc.c.h
@@ -19,7 +19,7 @@ main_heaps_kmalloc(size_t *size, u32 flags)
    ASSERT(kmalloc_initialized);
 
    // Iterate in reverse-order because the first heaps are the biggest ones.
-   for (int i = (int)used_heaps - 1; i >= 0; i--) {
+   for (int i = used_heaps - 1; i >= 0; i--) {
 
       ASSERT(heaps[i] != NULL);
 
@@ -55,7 +55,7 @@ main_heaps_kfree(void *ptr, size_t *size, u32 flags)
    const ulong vaddr = (ulong) ptr;
    ASSERT(kmalloc_initialized);
 
-   for (int i = (int)used_heaps - 1; i >= 0; i--) {
+   for (int i = used_heaps - 1; i >= 0; i--) {
 
       const ulong hva = heaps[i]->vaddr;
       const ulong hend = heaps[i]->heap_last_byte-heaps[i]->min_block_size+1;
